Picks next best natural size based on size
User can still provide their own, but it calculates it if it is not
provided. The pattern for near perfect identicons starting with 10 are:
10, 17, 22, 29, 34, 41, etc. So every 2 sizes are 12 apart. I add 12
until we pass the provided $scope.size and subtract 5 to try to use the
lower value if we can.

The lowest natural size this is configured for is 17px. Theoretically if
someone wanted one smaller than 10px, 10px would be a better natural
size, but I assumed more people would be in the range of 24px minimum so
it just cuts off an extra loop just in case (likely not even needed).
But either way, I believe 15px is the minimum for identicons to even be
useful.

Also added one test in case you ever set up the framework (it doesn't
appear to be set up right now).